+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.4.8
 - Fix missing Chicken Pool and Missing Pool warnings in Pumpkin Patch Village
 - Add some Missing Block and Item Tags
+- Fix Witch Hazel Blossom and Branch not being compostable and flammable
 
 # 2.4.7
 - Add Pale Pumpkin Seeds to Chicken and Parrot Food